#3 - Support serialization on dtos and add a new event on frame
