+++ hkb_core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod dtos;
 pub mod logger;
 
 #[cfg(any(feature = "sqlite-database", feature = "mysql-database",))]
